now i can find a cube from the kernel

+++ src/render.cpp
@@ -31,46 +31,30 @@ bool Render::init(device_t device)
 	_ccc = 0;
 	_pvpH = 0;
 	_pvpW = 0;
-	_chunk = 0;
 
 	if (cudaSuccess != cudaSetDevice(_device))
 	{
 		std::cerr<<"Render, cudaSetDevice error: "<<cudaGetErrorString(cudaGetLastError())<<std::endl;
 		return false;
 	}
-
+	if (cudaSuccess != cudaStreamCreate(&_stream))
+	{
+		std::cerr<<"Render, cuda create stream error: "<<cudaGetErrorString(cudaGetLastError())<<std::endl;
+		return false;
+	}
 	return true;
 }
 
 void Render::destroy()
 {
-	workpackage_t workP;
-	workP.work[0] = STOP;
-	// SEND STOP
-	for(int i=0; i<MAX_WORKERS; i++)
+	if (cudaSuccess != cudaStreamDestroy(_stream))
 	{
-		_octreeQueue[i].cond.lock();
-		_octreeQueue[i].queue.push(workP);
-		if (_octreeQueue[i].queue.size() == 1)
-			_octreeQueue[i].cond.signal();
-		_octreeQueue[i].cond.unlock();
+		std::cerr<<"Render, cuda create stream error: "<<cudaGetErrorString(cudaGetLastError())<<std::endl;
+		return;
 	}
-	// WAIT FOR STOP
-	_masterQueue.cond.lock();
-	for(int i=0; i<MAX_WORKERS; i++)
-	{
-		if (_masterQueue.queue.empty())
-			_masterQueue.cond.wait();
-		_masterQueue.queue.pop();
-	}
-	_masterQueue.cond.unlock();
-	for(int i=0; i<MAX_WORKERS; i++)
-	{
-		_octreeWorkers[i].join();
-		_cacheWorkers[i].join();
-		_poperWorkers[i].join();
-	}
+
 	_destroyVisibleCubes();
+	_cache.destroy();
 }
 
 void Render::_destroyVisibleCubes()
@@ -89,63 +73,17 @@ bool Render::setCache(ControlCubeCache * ccc)
 {
 	_ccc = ccc;	
 
-	if (_octree != 0 && _colors.r != 0 && _ccc != 0)
-	{
-		for(int i=0; i<MAX_WORKERS; i++)
-		{
-			_octreeWorkers[i].init(_octree, &_colors, _device, &_octreeQueue[i], &_cacheQueue[i], &_parameters, &_masterQueue, &_popQueue[i]);
-			_cacheWorkers[i].init(_octree, _ccc, _device, &_cacheQueue[i], &_octreeQueue[i], &_parameters);
-			_poperWorkers[i].init(&_popQueue[i], _cacheWorkers[i].getCache());
-		}
-		for(int i=0; i<MAX_WORKERS; i++)
-		{
-			_octreeWorkers[i].start();
-			_cacheWorkers[i].start();
-			_poperWorkers[i].start();
-		}
-	}
-
-	return true;
+	return _cache.init(_ccc);
 }
 
 void Render::setOctree(Octree * octree)
 { 
 	_octree = octree;
-	if (_octree != 0 && _colors.r != 0 && _ccc != 0)
-	{
-		for(int i=0; i<MAX_WORKERS; i++)
-		{
-			_octreeWorkers[i].init(_octree, &_colors, _device, &_octreeQueue[i], &_cacheQueue[i], &_parameters, &_masterQueue, &_popQueue[i]);
-			_cacheWorkers[i].init(_octree, _ccc, _device, &_cacheQueue[i], &_octreeQueue[i], &_parameters);
-			_poperWorkers[i].init(&_popQueue[i], _cacheWorkers[i].getCache());
-		}
-		for(int i=0; i<MAX_WORKERS; i++)
-		{
-			_octreeWorkers[i].start();
-			_cacheWorkers[i].start();
-			_poperWorkers[i].start();
-		}
-	}
 }
 
 void Render::setColors(color_t colors)
 { 
 	_colors = colors;
-	if (_octree != 0 && _colors.r != 0 && _ccc != 0)
-	{
-		for(int i=0; i<MAX_WORKERS; i++)
-		{
-			_octreeWorkers[i].init(_octree, &_colors, _device, &_octreeQueue[i], &_cacheQueue[i], &_parameters, &_masterQueue, &_popQueue[i]);
-			_cacheWorkers[i].init(_octree, _ccc, _device, &_cacheQueue[i], &_octreeQueue[i], &_parameters);
-			_poperWorkers[i].init(&_popQueue[i], _cacheWorkers[i].getCache());
-		}
-		for(int i=0; i<MAX_WORKERS; i++)
-		{
-			_octreeWorkers[i].start();
-			_cacheWorkers[i].start();
-			_poperWorkers[i].start();
-		}
-	}
 }
 
 bool Render::setViewPort(int pvpW, int pvpH)
@@ -166,10 +104,6 @@ bool Render::setViewPort(int pvpW, int pvpH)
 				return false;
 			}
 		}
-		if (_pvpH < 100)
-			_chunk = _pvpH;
-		else
-			_chunk = 100;
 	}
 
 	return true;
@@ -203,88 +137,59 @@ bool Render::_draw(	vmml::vector<4, float> origin, vmml::vector<4, float> LB,
 	clock.reset();
 	#endif
 
-	/* SET SHARED PARAMETERS */
-	_parameters.size = _pvpW*_pvpH;
-	_parameters.visibleCubes = _visibleCubes;
-	_parameters.visibleCubesGPU = _visibleCubesGPU;
-	_parameters.pvpW = _pvpW;
-	_parameters.pvpH = _pvpH;
-	_parameters.w = w;
-	_parameters.h = h;
-	_parameters.origin = VectorToFloat3(origin);
-	_parameters.LB = VectorToFloat3(LB);
-	_parameters.up = VectorToFloat3(up);
-	_parameters.right = VectorToFloat3(right);
-	_parameters.pixelBuffer = _pixelBuffer;
-	/* END SET SHARED PARAMETERS*/
-	
-	workpackage_t workP;
-	workP.work[0] = START_FRAME;
-	/* SEND START FRAME */
-	for(int i=0; i<MAX_WORKERS; i++)
+	bool newIteration = true;
+
+	int iterations = 0;
+	while(newIteration)
 	{
-		_octreeQueue[i].cond.lock();
-		_octreeQueue[i].queue.push(workP);
-		if (_octreeQueue[i].queue.size() == 1)
-			_octreeQueue[i].cond.signal();
-		_octreeQueue[i].cond.unlock();
-	}
+		float ** tCubes = _cache.syncAndGetTableCubes(_stream);
 
-	_masterQueue.cond.lock();
+		getBoxIntersectedOctree(_octree->getOctree(), _octree->getSizes(), _octree->getnLevels(),
+								VectorToFloat3(origin), VectorToFloat3(LB), VectorToFloat3(up), VectorToFloat3(right),
+								w, h, _pvpW, _pvpH, _octree->getRayCastingLevel(), _octree->getCubeLevel(), _pvpW*_pvpH, 
+								_visibleCubesGPU, 0, (VectorToInt3(_octree->getRealDim())), 
+								_colors.r, _colors.g, _colors.b, _pixelBuffer, 
+								_octree->getIsosurface(), _octree->getMaxHeight(), 
+								tCubes, _stream);
 
-	/* SEND FRAME */
-	workP.work[0] = FRAME;
-	workP.work[1] = 0; 
-	workP.work[2] = _chunk * _pvpW;
-
-	int p = 0;
-	bool notMulti = _pvpH % _chunk != 0;
-	int works = _pvpH / _chunk;
-	while(p < works)
-	{
-		for(int i=0; i<MAX_WORKERS && p < works; i++)
+		if (cudaSuccess != cudaMemcpyAsync((void*)(_visibleCubes), (void*)(_visibleCubesGPU), _pvpH*_pvpW*sizeof(visibleCube_t), cudaMemcpyDeviceToHost, _stream))
 		{
-			_octreeQueue[i].cond.lock();
-			_octreeQueue[i].queue.push(workP);
-			if (_octreeQueue[i].queue.size() == 1)
-				_octreeQueue[i].cond.signal();
-			_octreeQueue[i].cond.unlock();
-		
-			workP.work[1] += _chunk * _pvpW;
-			p++;
+			std::cerr<<"Visible cubes, error updating cpu copy: "<<cudaGetErrorString(cudaGetLastError())<<std::endl;
+			return false;
 		}
-	}
-	if (notMulti)
-	{
-		workP.work[1] =  works * _chunk * _pvpW; 
-		workP.work[2] = (_pvpH % _chunk) * _pvpW;
-		works++;
-		_octreeQueue[0].cond.lock();
-		_octreeQueue[0].queue.push(workP);
-		if (_octreeQueue[0].queue.size() == 1)
-			_octreeQueue[0].cond.signal();
-		_octreeQueue[0].cond.unlock();
-	}
+		if (cudaSuccess != cudaStreamSynchronize(_stream))
+		{
+			std::cerr<<"Error sync: "<<cudaGetErrorString(cudaGetLastError())<<std::endl;
+			return false;
+		}
 
-	// WAIT FOR FRAMES
-	for(int i=0; i<works; i++)
-	{
-		if (_masterQueue.queue.empty())
-			_masterQueue.cond.wait();
-		//workpackage_t p = _masterQueue.queue.front();
-		_masterQueue.queue.pop();
-	}
-	_masterQueue.cond.unlock();
+		int cF = 0;
+		for(int i=0; i<_pvpH*_pvpW; i++)
+		{
+			if (_visibleCubes[i].idCube != 0)
+				_cache.popCubes(_visibleCubes[i].idCube);
 
-	/* SEND FINIDH FRAME */
-	workP.work[0] = FINISH_FRAME;
-	for(int i=0; i<MAX_WORKERS; i++)
-	{
-		_octreeQueue[i].cond.lock();
-		_octreeQueue[i].queue.push(workP);
-		if (_octreeQueue[i].queue.size() == 1)
-			_octreeQueue[i].cond.signal();
-		_octreeQueue[i].cond.unlock();
+			_cache.pushCubes(&_visibleCubes[i]);
+
+			if(_visibleCubes[i].state == DONE || _visibleCubes[i].state == PAINTED)
+			{
+				cF++;
+			}
+		}
+
+		if (cF == _pvpH*_pvpW)
+		{
+			newIteration = false;
+		}
+		else
+		{
+			if (cudaSuccess != cudaMemcpyAsync((void*)(_visibleCubesGPU), (void*)(_visibleCubes), _pvpH*_pvpW*sizeof(visibleCube_t), cudaMemcpyHostToDevice, _stream))
+			{
+				std::cerr<<"Visible cubes, error updating gpu copy: "<<cudaGetErrorString(cudaGetLastError())<<std::endl;
+				return false;
+			}
+		}
+		iterations++;
 	}
 
 	#ifdef TIMING
@@ -305,12 +210,15 @@ bool Render::frameDraw(	vmml::vector<4, float> origin, vmml::vector<4, float> LB
 	}
 	else
 	{
+		_cache.setRayCastingLevel(_octree->getRayCastingLevel());
+		_cache.startFrame();
 		if (cudaSuccess != cudaMemset((void**)_visibleCubesGPU, 0, _size*sizeof(visibleCube_t)))
 		{
 			std::cerr<<"Render, error int memory: "<<cudaGetErrorString(cudaGetLastError())<<std::endl;
 			return false;
 		}
 		result =  _draw(origin, LB, up, right, w, h);
+		_cache.finishFrame();
 	}
 
 	return result;
